+++ src/main/java/com/company/calculator/services/ArithmeticCalculatorServiceImpl.java
@@ -2,14 +2,14 @@ package com.company.calculator.services;
 
 import org.springframework.stereotype.Service;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Service
 public class ArithmeticCalculatorServiceImpl implements ArithmeticCalculatorService {
 
     @Override
-    public BigInteger addition(final BigInteger firstTerm,
-                               final BigInteger secondTerm) {
+    public BigDecimal addition(final BigDecimal firstTerm,
+                               final BigDecimal secondTerm) {
         return firstTerm.add(secondTerm);
     }
 }
